NS8 Protect Connector for SAP CX 0.7.0

+++ connector/ns8backoffice/backoffice/src/com/ns8/hybris/backoffice/listview/renderer/Ns8ScoreRenderer.java
@@ -49,7 +49,7 @@ public class Ns8ScoreRenderer extends AbstractWidgetComponentRenderer<Listcell, 
             } else {
                 final Double amount = (Double) propertyValueService.readValue(object, SCORE_KEY);
                 if (amount != null && permissionFacade.canReadProperty(dataType.getCode(), SCORE_KEY)) {
-                    listcell.setLabel(String.valueOf(amount));
+                    listcell.setLabel(String.valueOf(amount.intValue()));
                 } else {
                     listcell.setLabel(NOT_AVAILABLE_MESSAGE);
                 }
